Extract default Cash on Delivery model creation
Sharing the defaultPayInPersonGateway, currently used to enable Cash on Delivery from the IPP Onboarding screen, will allow us to enable it from the IPP Menu toggle as well.
